Rewrote the formatting
Instead of having "writers" that wrap a specific stream, there are now
"formatters" that return strings or streams thereof.
This makes things easier to compose (for one thing, the syntax for
function pointers can now be generated correctly, with the return type
formatted first, but emitted after the parameter types).