Ensure numeric values are locale-insensitive
This adds `Literal` methods for all the builtin numeric types (except
for `nint` and `nuint` who can only be initialized via `int` and `uint`,
respectively, anyway).
These ensure the invariant culture is used for formatting and now
explicitly disambiguate between all the types via either the applicable
suffix or an explicit cast (so `-7UL` and `(sbyte) -7` instead of just
`-7`).