Avoid a crash for an unresolved type for a value
The code that tried to handle enum values assumed the value's type would
always be resolvable.
However, Pipelines.Sockets.Unofficial 2.1.16 included a case where a
default value is provided for a parameter of type
`System.Threading.ThreadPriority`, which apparently cannot be resolved,
resulting in a `NullReferenceException`.
This no longer crashes; instead the value will be rendered normally,
i.e. as a plain integer.