Use configuration for attribute inclusion
Instead of having runtime logic that decides which attributes are to be
considered relevant to the API, there are now command-line arguments
to include/exclude particular attribute names.

In the normal use case, these are provided through the
`ApiReferenceIncludeAttribute` and `ApiReferenceExcludeAttribute`
item groups in the project file.

By default, the `ApiReferenceIncludeAttribute` item group will be empty,
selecting all attributes for inclusion, and the
`ApiReferenceExcludeAttribute` group gets a short list of
definitely-not-API-related attributes.