Fix the attribute inclusion/exclusion handling
The MSBuild item groups don't actually enable full wildcard support,
because they will often trigger matching them against the filesystem.
So drop the documentation saying that wildcards are supported.

In addition, the MSBuild code setting up the `-ia` and `-ea` parameters
was incorrect, causing none of the configured attributes to be passed.