Handle special attributes
Some attributes are not compiled to custom attributes but are used by
the compiler to generate specific IL. The resulting flags are now
formatted back as attributes (unconditionally - because these impact the
IL signatures, they are considered relevant for public API).

This processing is currently limited to `[Serializable]` on types, and
`[MethodImpl(MethodImplOptions)]` on methods.