Fix method sort order
With .NET 8, the sorting applied for methods results in a different
order than on older targets.
Instead of a `SortedDictionary` with a custom-created sort key, we
now use a `SortedSet` and provide an implementation of `IComparer` for
method definitions. This allows finer control over the comparison.

Note that this now also causes the sorting for parameter types to be
based on their text form in the selected formatter. This means that
`System.Int16` (formatted as `short`) now sorts after `System.Int32`
(formatted as `int`).

The same approach should probably be applied to all other places where
we currently use `SortedDictionary`.

Fixes #39.