Add initial `NumericIntPtr` support
This handles the fact that .NET 7 and later no longer add
`[NativeIntegerAttribute]` to `[U]IntPtr` references that should be
considered `n[u]int`.

Currently this uses a version check; that should be changed at some
point to check for the actual runtime feature, but it seems like doing
that with Cecil is not easy (it wants to resolve things to the active
runtime, not the runtime referenced by the assembly being processed).