Adjust method sorting
Methods marked as special-named (i.e. mostly operators) included the
return type in the method name, which affects sorting.
This has been adjusted so that the return type is only taken into
account when sorting if all other components (method name, generic
types, and parameters) are all the same (which probably should not ever
be the case).