Adjust C# enum field generation
There is specific handling for enums based on `ushort` where all values
can be reasonably represented as a character, in which case all fields
will use a character literal.

The general case was improved as well, ensuring no casts or suffixes are
used for values of enum fields when the enum is not based on `int`.