Document the tool project
The tool project now uses `dotnet-api-reference` as both the assembly
and command name. The root namespace has been changed to match the
project name.

Note that this also reduces the set of values supported by the `-eh`
option as well as the `-f` option (with a corresponding change to what
can be specified for `ApiReferenceFormat` in MSBuild).

This usage text for the tool has been adjusted (slightly more
information is provided), and the README for the tool project now also
documents this usage.